+++ client/actions/TerminalActionCreators.js
@@ -1,35 +1,17 @@
 import _s from 'underscore.string';
 
+import {
+  COMMAND_DEFINITION,
+  COMMANDS,
+  MINIMIST_OPTIONS_FOR_COMMAND,
+  SHELL_INPUT_MODE_ALIASES,
+} from './commands';
 import ActionTypes from 'consts/ActionTypes';
-import ShellInputModes from 'consts/ShellInputModes';
-import AppModel from 'containers/AppModel';
 import { parse } from 'lib/command-parser';
-import { PLAYER_STATE_CODES } from 'models/GameModel';
 
 
-export function _selectShellInputMode(playerStateCode) {
-  return ShellInputModes[({
-    [PLAYER_STATE_CODES.ADVENTURING]: ShellInputModes.ADVENTURE,
-    [PLAYER_STATE_CODES.BATTLING]: ShellInputModes.BATTLE,
-  }[playerStateCode]) || ShellInputModes.DEFAULT];
-}
-
-
-const SHELL_INPUT_MODE_ALIASES = {
-  _test: [
-    [/ba/, 'ba_to_baz'],
-    [/baz/, ''],
-  ],
-
-  [ShellInputModes.WIZARD]: [
-    [/^adv/, '_wizard adventuring'],
-    [/^bat/, '_wizard battling'],
-  ],
-};
-
-export function _applyShellInputModeAliasesToInput(shellInputMode, input) {
-  const filters = SHELL_INPUT_MODE_ALIASES[shellInputMode] || [];
-  filters.some(([matcher, replacement]) => {
+export function _applyShellInputModeAliasesToInput(aliases, input) {
+  aliases.some(([matcher, replacement]) => {
     const replaced = input.replace(matcher, replacement);
     if (replaced !== input) {
       input = replaced;
@@ -39,100 +21,6 @@ export function _applyShellInputModeAliasesToInput(shellInputMode, input) {
   });
   return input;
 }
-
-
-const COMMAND_DEFINITION = {
-  commands: {
-    _wizard: {
-      default: 'on',
-      commands: {
-        adventuring: null,
-        battling: null,
-        off: null,
-        on: null,
-      },
-    },
-    alias: null,
-    character: {
-      default: 'index',
-      commands: {
-        index: null,
-        list: null,
-        select: null,
-        show: null,
-      }
-    },
-    config: null,
-    dictionary: null,
-    guild: null,
-    help: {
-      default: 'welcome',
-      commands: {
-        welcome: null,
-      }
-    },
-    item: {
-      default: 'index',
-      commands: {
-        index: null,
-        list: null,
-        show: null,
-      }
-    },
-    recruit: null,
-  },
-};
-
-const MINIMIST_OPTIONS_FOR_COMMAND = {
-};
-
-const COMMANDS = {
-
-  '_wizard-adventuring': function wizardAdventuring() {
-    const { game } = AppModel.getInstance();
-    game._isAdventuring = true;
-    game._isBattling = false;
-    return {
-      type: ActionTypes.APPLY_COMMAND_EXECUTION,
-      newShellInputMode: _selectShellInputMode(game.getPlayerStateCode()),
-    };
-  },
-
-  '_wizard-battling': function wizardBattling() {
-    const { game } = AppModel.getInstance();
-    game._isAdventuring = true;
-    game._isBattling = true;
-    return {
-      type: ActionTypes.APPLY_COMMAND_EXECUTION,
-      newShellInputMode: _selectShellInputMode(game.getPlayerStateCode()),
-    };
-  },
-
-  '_wizard-off': function wizardOn() {
-    return {
-      type: ActionTypes.APPLY_COMMAND_EXECUTION,
-      newShellInputMode: ShellInputModes.DEFAULT,
-    };
-  },
-
-  '_wizard-on': function wizardOn() {
-    return {
-      type: ActionTypes.APPLY_COMMAND_EXECUTION,
-      newShellInputMode: ShellInputModes.WIZARD,
-    };
-  },
-
-  'help-welcome': function helpIndex() {
-    return {
-      type: ActionTypes.APPLY_COMMAND_EXECUTION,
-      output: [
-        '{magenta-fg}HAKUSURA{/} - A text-based hack & slash RPG',
-        '',
-        'If you are a beginner, please execute the `{green-fg}tutorial{/}` command.',
-      ].join('\n')
-    };
-  },
-};
 
 
 const TerminalActionCreators = {
@@ -149,7 +37,10 @@ const TerminalActionCreators = {
 
   executeCommand(shellInputMode, rawInput) {
 
-    const input = _applyShellInputModeAliasesToInput(shellInputMode, rawInput);
+    let input = rawInput;
+    if (SHELL_INPUT_MODE_ALIASES[shellInputMode]) {
+      input = _applyShellInputModeAliasesToInput(SHELL_INPUT_MODE_ALIASES[shellInputMode], rawInput);
+    }
 
     if (_s.trim(input) === '') {
       return {
